fix: remove qf-eligibility + passport page
